+++ src/test/kotlin/no/nav/dagpenger/regel/api/v1/models/FaktumTest.kt
@@ -1,0 +1,11 @@
+package no.nav.dagpenger.regel.api.v1.models
+
+import org.junit.jupiter.api.Test
+
+
+internal class FaktumTest {
+    @Test
+    fun `Mapping from Packet to Faktum`() {
+
+    }
+}
