+++ src/test/kotlin/no/nav/dagpenger/regel/api/v1/models/SubsumsjonTest.kt
@@ -47,5 +47,4 @@ internal class SubsumsjonTest {
             bruktInntektsPeriode shouldBe InntektsPeriode(YearMonth.of(2019, 5), YearMonth.of(2019, 5))
         }
     }
-
 }
